src/image_classification/API/requirements.txt
aniso8601==7.0.0
certifi==2019.6.16
Click==7.0
Flask==1.1.1
Flask-RESTful==0.3.7
imutils==0.5.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.16.4
opencv-python==4.1.0.25
Pillow==6.1.0
pytesseract==0.2.7
pytz==2019.1
simplejson==3.16.0
six==1.12.0
Werkzeug==0.15.4
Flask-Testing==0.7.1
tensorflow
flask_cors
imutils==0.5.3
